+++ src/Animation.lua
@@ -1,65 +1,28 @@
 local AnimationConfig = require(script.Parent.AnimationConfig)
 local util = require(script.Parent.util)
+local helpers = require(script.Parent.helpers)
 
 local Animation = {}
 Animation.__index = Animation
 
 type animationType = number | UDim | UDim2 | Vector2 | Vector3 | Color3
 
-local function getValuesFromType(data)
-    local dataType = typeof(data)
-
-    if dataType == "number" then
-        return { data }
-    elseif dataType == "UDim" then
-        return { data.Scale, data.Offset }
-    elseif dataType == "UDim2" then
-        return { data.X.Scale, data.X.Offset, data.Y.Scale, data.Y.Offset }
-    elseif dataType == "Vector2" then
-        return { data.X, data.Y }
-    elseif dataType == "Vector3" then
-        return { data.X, data.Y, data.Z }
-    elseif dataType == "Color3" then
-        return { data.R, data.G, data.B }
-    end
-
-    error("Unsupported type: " .. dataType)
-end
-
-local function getTypeFromValues(type: string, values: { number })
-    if type == "number" then
-        return values[1]
-    elseif type == "UDim" then
-        return UDim.new(values[1], values[2])
-    elseif type == "UDim2" then
-        return UDim2.new(values[1], values[2], values[3], values[4])
-    elseif type == "Vector2" then
-        return Vector2.new(values[1], values[2])
-    elseif type == "Vector3" then
-        return Vector3.new(values[1], values[2], values[3])
-    elseif type == "Color3" then
-        return Color3.new(values[1], values[2], values[3])
-    end
-
-    error("Unsupported type: " .. type)
-end
-
 function Animation.new(props)
     if props.to and props.from then
         assert(typeof(props.to) == typeof(props.from), "`to` and `from` must be the same type")
     end
-    local length = #getValuesFromType(props.from or props.to)
+    local length = #helpers.getValuesFromType(props.from or props.to)
 
 	return setmetatable({
-        values = getValuesFromType(props.from or props.to),
-        toValues = getValuesFromType(props.to or props.from),
-        fromValues = getValuesFromType(props.from or props.to),
+        values = helpers.getValuesFromType(props.from or props.to),
+        toValues = helpers.getValuesFromType(props.to or props.from),
+        fromValues = helpers.getValuesFromType(props.from or props.to),
         type = typeof(props.from or props.to),
         config = AnimationConfig:mergeConfig(props.config or {}),
         immediate = props.immediate,
 
         v0 = table.create(length, nil),
-        lastPosition = getValuesFromType(props.from or props.to),
+        lastPosition = helpers.getValuesFromType(props.from or props.to),
         lastVelocity = table.create(length, nil),
         done = table.create(length, false),
         elapsedTime = table.create(length, 0),
@@ -79,9 +42,9 @@ end
 
 function Animation:setProps(props)
     if props then
-        self.toValues = if props.to then getValuesFromType(props.to) else self.toValues
-        self.lastPosition = if props.from then getValuesFromType(props.from) else self.lastPosition
-        self.fromValues = if props.from then getValuesFromType(props.from) else util.copy(self.lastPosition)
+        self.toValues = if props.to then helpers.getValuesFromType(props.to) else self.toValues
+        self.lastPosition = if props.from then helpers.getValuesFromType(props.from) else self.lastPosition
+        self.fromValues = if props.from then helpers.getValuesFromType(props.from) else util.copy(self.lastPosition)
         self.config = AnimationConfig:mergeConfig(props.config or {})
         self.immediate = if props.immediate ~= nil then props.immediate else self.immediate
 
@@ -92,7 +55,7 @@ function Animation:setProps(props)
 end
 
 function Animation:getValue()
-    return getTypeFromValues(self.type, self.values)
+    return helpers.getTypeFromValues(self.type, self.values)
 end
 
 function Animation:stop()