+++ src/init.lua
@@ -1,7 +1,7 @@
 local RoactSpring = {
-	useSpring = require(script.useSpring),
-	useSprings = require(script.useSprings),
-	useTrail = require(script.useTrail),
+	useSpring = require(script.hooks.useSpring),
+	useSprings = require(script.hooks.useSprings),
+	useTrail = require(script.hooks.useTrail),
 	Controller = require(script.Controller),
 	SpringValue = require(script.SpringValue),
 	config = require(script.constants).config,