+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added `progress` config for easing animations ([@chriscerie](https://github.com/chriscerie) in [#13](https://github.com/chriscerie/roact-spring/pull/13))
+* Hooks now cancel animations when they are unmounted
 
 ## 0.2.1 (Feburary 17, 2022)